refactor(cli): split pi-transcript and pi-tui-runner into focused modules
Behavior-preserving decomposition of the two largest CLI TUI files; no
runtime logic or test assertions change (import paths only).

pi-transcript.ts (1423 -> ~770 lines) splits into:
- pi-transcript-format.ts: leaf formatting helpers (fitLine, renderIndented,
  formatToolResultContent, formatUnknown(Inline), collapseToSingleLine,
  limitText, markdownTheme) with no transcript-type deps.
- pi-transcript-tools.ts: the tool-card render subsystem (renderToolBlock and
  its compact/expanded helpers), importing the format leaf and type-only tool
  entry types.
pi-transcript.ts keeps state, ingestion, high-level render, and the non-tool
block renderers. No public export moved, so index.ts and the test file are
untouched.

pi-tui-runner.ts (1239 -> ~850 lines) splits into:
- pi-tui-layout.ts: presentational components (transcript, status line, the
  scroll/anchor Container) plus anchor helpers.
- pi-tui-pickers.ts: autocomplete provider, picker overlay, and picker item
  builders.
runMakaPiTui and its input type stay in the runner and import the two modules.

typecheck clean; all 190 tests green.
